+++ src/components/04-templates/news/news-list/news-list.config.js
@@ -50,4 +50,18 @@ module.exports = {
         ],
         teasers: teaserItemData,
     },
+    variants: [
+        {
+            name: 'projets',
+            context: {
+                pageTitle: 'Projets',
+                breadcrumb: [
+                    {
+                        page: 'vd.ch',
+                        url: '#',
+                    },
+                ],
+            },
+        },
+    ],
 };