+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Rename `@acceuil-securise` in `@tableau-de-bord` and change content
 
 ### Added
+- Add `@news-list--projets` template
 - Add `@search-form--regular-option`component
 
 ### Fix